quant_finance/requirements.txt
flask
flask-cors
requests
numpy
pandas
python-dotenv
prophet
statsmodels
scikit-learn
tensorflow